+++ login-workflow/src/components/password/ChangePasswordModal.tsx
@@ -11,7 +11,17 @@ import {
     useInjectedUIContext,
     AccountActions,
 } from '@pxblue/react-auth-shared';
-import { Dialog, DialogTitle, DialogActions, DialogContent, Button, Grid, useTheme, Divider } from '@material-ui/core';
+import {
+    Dialog,
+    DialogTitle,
+    DialogActions,
+    DialogContent,
+    Button,
+    Grid,
+    useTheme,
+    Divider,
+    useMediaQuery,
+} from '@material-ui/core';
 import { ChangePasswordForm } from './ChangePasswordForm';
 import { SecureTextField } from '../SecureTextField';
 import { SimpleDialog } from '../SimpleDialog';
@@ -35,6 +45,7 @@ export const ChangePasswordModal: React.FC = () => {
     const securityHelper = useSecurityActions();
     const theme = useTheme();
     const sharedClasses = useDialogStyles();
+    const matchesXS = useMediaQuery(theme.breakpoints.down('xs'));
 
     const passwordRef = useRef(null);
     const confirmRef = useRef(null);
@@ -82,14 +93,14 @@ export const ChangePasswordModal: React.FC = () => {
         body = (
             <FinishState
                 icon={<CheckCircle color={'primary'} style={{ fontSize: 100, marginBottom: theme.spacing(2) }} />}
-                title={t('CHANGE_PASSWORD.PASSWORD_CHANGED')}
-                description={t('CHANGE_PASSWORD.SUCCESS_MESSAGE')}
+                title={t('pxb:CHANGE_PASSWORD.PASSWORD_CHANGED')}
+                description={t('pxb:CHANGE_PASSWORD.SUCCESS_MESSAGE')}
             />
         );
     } else {
         body = (
             <ChangePasswordForm
-                passwordLabel={t('LABELS.NEW_PASSWORD')}
+                passwordLabel={t('pxb:LABELS.NEW_PASSWORD')}
                 onPasswordChange={updateFields}
                 passwordRef={passwordRef}
                 confirmRef={confirmRef}
@@ -104,7 +115,7 @@ export const ChangePasswordModal: React.FC = () => {
             >
                 <SecureTextField
                     id="current-password"
-                    label={t('LABELS.CURRENT_PASSWORD')}
+                    label={t('pxb:LABELS.CURRENT_PASSWORD')}
                     value={currentPassword}
                     onChange={(evt: ChangeEvent<HTMLInputElement>): void => setCurrentPassword(evt.target.value)}
                     onKeyPress={(e): void => {
@@ -119,7 +130,7 @@ export const ChangePasswordModal: React.FC = () => {
 
     const errorDialog = (
         <SimpleDialog
-            title={t('MESSAGES.ERROR')}
+            title={t('pxb:MESSAGES.ERROR')}
             body={transitState.transitErrorMessage ?? ''}
             open={transitState.transitErrorMessage !== null && !hasAcknowledgedError}
             onClose={(): void => {
@@ -136,9 +147,14 @@ export const ChangePasswordModal: React.FC = () => {
     }, [setTransitState, setCurrentPassword, setPassword, setConfirm]);
 
     return (
-        <Dialog open={securityState.isShowingChangePassword} maxWidth={'xs'} onExited={resetForm}>
+        <Dialog
+            fullScreen={matchesXS ? true : false}
+            open={securityState.isShowingChangePassword}
+            maxWidth={'xs'}
+            onExited={resetForm}
+        >
             {errorDialog}
-            <DialogTitle className={sharedClasses.dialogTitle}>{t('CHANGE_PASSWORD.PASSWORD')}</DialogTitle>
+            <DialogTitle className={sharedClasses.dialogTitle}>{t('pxb:CHANGE_PASSWORD.PASSWORD')}</DialogTitle>
             <DialogContent className={sharedClasses.dialogContent} style={{ flex: '1 1 auto' }}>
                 {body}
             </DialogContent>
@@ -152,7 +168,7 @@ export const ChangePasswordModal: React.FC = () => {
                             className={sharedClasses.dialogButton}
                             onClick={(): void => securityHelper.hideChangePassword()}
                         >
-                            {t('ACTIONS.BACK')}
+                            {t('pxb:ACTIONS.BACK')}
                         </Button>
                     )}
                     <Button
@@ -173,7 +189,7 @@ export const ChangePasswordModal: React.FC = () => {
                                 : changePassword
                         }
                     >
-                        {success ? t('ACTIONS.LOG_IN') : t('ACTIONS.OKAY')}
+                        {success ? t('pxb:ACTIONS.LOG_IN') : t('pxb:ACTIONS.OKAY')}
                     </Button>
                 </Grid>
             </DialogActions>