+++ login-workflow/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v2.2.0 (June 28, 2021)
+
+### Changed
+
+-   Updated styles to improve mobile responsiveness.
+-   Updated styles to improve dark theme.
+
+### Added
+
+-   Ability to use a username or email input field via the `loginType` prop on the `AuthUIContextProvider`.
+-   Support for Spanish language.
+-   Support for Chinese language.
+-   Login screen email input validation error message.
+
 ## v2.1.0 (April 30, 2021)
 
 ### Added